requirements-render.txt
Pillow
flask
flask-cors
boto3
python-dotenv
gunicorn